+++ vllm_gguf_plugin/triton/gemm/iq_quant/iq1_s.py
@@ -3,11 +3,11 @@ import triton
 import triton.language as tl
 
 from .iq_tables import get_iq_table_tensors
-from .utils import GGML_TYPE_IQ2_XS, load_f16_from_u8, load_u16_from_u8, load_x_chunk, run_triton_kernel
+from ..utils import GGML_TYPE_IQ1_S, load_f16_from_u8, load_u16_from_u8, load_x_chunk, run_triton_kernel
 
 
 @triton.jit
-def iq2_xs_gemm_kernel(
+def iq1_s_gemm_kernel(
     x_ptr,
     w_u8_ptr,
     y_ptr,
@@ -20,7 +20,6 @@ def iq2_xs_gemm_kernel(
     stride_ym,
     stride_yn,
     grid_ptr,
-    sign_ptr,
     BLOCK_M: tl.constexpr,
     BLOCK_N: tl.constexpr,
     BLOCK_K_BLOCKS: tl.constexpr,
@@ -30,21 +29,20 @@ def iq2_xs_gemm_kernel(
     offs_m = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
     offs_n = pid_n * BLOCK_N + tl.arange(0, BLOCK_N)
     offs_8 = tl.arange(0, 8)
-    sign_mask = (1 << offs_8).to(tl.uint8)
     n_mask = offs_n < n
 
     acc = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
     w_row_ptrs = w_u8_ptr + offs_n * stride_wn
 
     for kb in range(0, num_k_blocks):
-        block_ptrs = w_row_ptrs + kb * 74
+        block_ptrs = w_row_ptrs + kb * 50
 
         for ib in range(8):
-            scale_byte = tl.load(block_ptrs + 66 + ib, mask=n_mask, other=0)
+            qh = load_u16_from_u8(block_ptrs + 34 + 2 * ib, n_mask)
+            delta_num = tl.where((qh & 0x8000) != 0, -9, -7).to(tl.int16)
             for il in range(4):
-                q2 = load_u16_from_u8(block_ptrs + 2 + 2 * (4 * ib + il), n_mask).to(tl.int32)
-                grid_idx = q2 & 0x1FF
-                signs = tl.load(sign_ptr + (q2 >> 9), mask=n_mask, other=0)
+                idx = tl.load(block_ptrs + 2 + 4 * ib + il, mask=n_mask, other=0).to(tl.int32)
+                idx = idx | ((((qh >> (3 * il)) & 0x7).to(tl.int32)) << 8)
                 x_tile = load_x_chunk(
                     x_ptr,
                     stride_xm,
@@ -56,14 +54,13 @@ def iq2_xs_gemm_kernel(
                 )
                 x_dtype = x_tile.dtype
                 d = load_f16_from_u8(block_ptrs + 0, n_mask).to(x_dtype)
-                scale = (((scale_byte >> (4 * (il // 2))) & 0x0F).to(x_dtype) + 0.5) * 0.25
+                scale = (d * (2 * ((qh >> 12) & 0x7).to(x_dtype) + 1.0) * 0.125).to(x_dtype)
                 grid = tl.load(
-                    grid_ptr + grid_idx[:, None] * 8 + offs_8[None, :],
+                    grid_ptr + idx[:, None] * 8 + offs_8[None, :],
                     mask=n_mask[:, None],
                     other=0,
                 ).to(x_dtype)
-                sign = tl.where((signs[:, None] & sign_mask[None, :]) != 0, -1, 1).to(x_dtype)
-                q_tile = (grid * sign * (d * scale).to(x_dtype)[:, None]).to(x_dtype)
+                q_tile = ((((grid.to(tl.int16) * 8) + delta_num[:, None]).to(x_dtype)) * scale[:, None]).to(x_dtype)
                 acc += tl.sum(x_tile[:, None, :] * q_tile[None, :, :], axis=2)
 
     y_ptrs = y_ptr + offs_m[:, None] * stride_ym + offs_n[None, :] * stride_yn
@@ -71,13 +68,13 @@ def iq2_xs_gemm_kernel(
     tl.store(y_ptrs, acc, mask=y_mask)
 
 
-def ggml_gemm_iq2_xs_triton(W: torch.Tensor, X: torch.Tensor, row: int) -> torch.Tensor:
+def ggml_gemm_iq1_s_triton(W: torch.Tensor, X: torch.Tensor, row: int) -> torch.Tensor:
     tables = get_iq_table_tensors(W.device)
     return run_triton_kernel(
-        iq2_xs_gemm_kernel,
+        iq1_s_gemm_kernel,
         W,
         X,
         row,
-        GGML_TYPE_IQ2_XS,
-        extra_args=(tables["iq2xs_grid"], tables["ksigns_iq2xs"]),
+        GGML_TYPE_IQ1_S,
+        extra_args=(tables["iq1s_grid"],),
     )
