+++ vllm_gguf_plugin/triton/gemm/k_quant/q2_k.py
@@ -2,7 +2,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import GGML_TYPE_Q2_K, load_f16_from_u8, load_x_chunk, run_triton_kernel
+from ..utils import GGML_TYPE_Q2_K, load_f16_from_u8, load_x_chunk, run_triton_kernel
 
 
 @triton.jit
