+++ vllm_gguf_plugin/triton/gemm/standard_quant/q5_0.py
@@ -2,7 +2,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import (
+from ..utils import (
     GGML_TYPE_Q5_0,
     load_f16_from_u8,
     load_u32_from_u8,
