+++ vllm_gguf_plugin/triton/gemm/standard_quant/q8_1.py
@@ -2,7 +2,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import GGML_TYPE_Q8_1, load_f16_from_u8, load_x_tile, run_triton_kernel
+from ..utils import GGML_TYPE_Q8_1, load_f16_from_u8, load_x_tile, run_triton_kernel
 
 
 @triton.jit
